+++ src/types/currencies.ts
@@ -1,4 +1,3 @@
-import {UserReadSerializer} from 'types/api/users';
 import {CreatedModified} from 'types/createdModified';
 import {Dict} from 'types/generic';
 
@@ -7,14 +6,6 @@ export interface Currency extends CreatedModified {
   id: number;
   logo: string;
   owner: number;
-  ticker: string;
-}
-
-export interface CurrencyReadDetailSerializer extends CreatedModified {
-  domain: string | null;
-  id: number;
-  logo: string;
-  owner: UserReadSerializer;
   ticker: string;
 }
